% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-prLogistic.R
\name{coef.prLogistic}
\alias{coef.prLogistic}
\title{Extract prevalence ratio point estimates}
\usage{
\method{coef}{prLogistic}(object, ...)
}
\arguments{
\item{object}{A \code{prLogistic} object.}

\item{...}{Currently ignored.}
}
\value{
Named numeric vector of PR estimates.
}
\description{
Extract prevalence ratio point estimates
}
