% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-prLogistic.R
\name{confint.prLogistic}
\alias{confint.prLogistic}
\title{Extract confidence intervals for prevalence ratios}
\usage{
\method{confint}{prLogistic}(object, parm, level, type = "percentile", ...)
}
\arguments{
\item{object}{A \code{prLogistic} object.}

\item{parm}{Ignored (all parameters returned).}

\item{level}{Ignored (level is stored in the object).}

\item{type}{For bootstrap objects: \code{"normal"} or \code{"percentile"}.}

\item{...}{Currently ignored.}
}
\value{
Numeric matrix with lower and upper bounds.
}
\description{
Extract confidence intervals for prevalence ratios
}
