% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{downer}
\alias{downer}
\title{Downer Cow Survival Data}
\format{
A data frame with 216 rows and 5 variables:
\describe{
\item{AST}{Aspartate aminotransferase (enzyme marker): 0 = normal,
1 = elevated.}
\item{CK}{Creatine kinase (enzyme marker): 0 = normal, 1 = elevated.}
\item{Calving}{Whether the downer condition was related to calving:
0 = No, 1 = Yes.}
\item{Myopathy}{Presence of myopathy: factor with levels \code{"No"}, \code{"Yes"}.}
\item{Survival}{Outcome: factor with levels \code{"Died"}, \code{"Survived"}.
Binary outcome.}
}
}
\source{
Dohoo, I., Martin, W. & Stryhn, H. (2003). \emph{Veterinary Epidemiologic
Research}. AVC Inc., Prince Edward Island, Canada.
}
\usage{
downer
}
\description{
Veterinary study of downer cows (cattle unable to rise after calving).
The outcome is whether the animal survived to discharge.
}
\examples{
data(downer)
prop.table(table(downer$Survival))

fit <- glm(as.integer(Survival == "Survived") ~ Myopathy + AST + CK + Calving,
           family = binomial, data = downer)
prLogisticDelta(fit)
}
\keyword{datasets}
