\name{notdocumentedyet}
\alias{notdocumentedyet}


\alias{boot.conditional}
\alias{boot.marg}
\alias{expandSlash} 
\alias{findbars} 
\alias{is.glm} 
\alias{is.glmer}
\alias{makeInteraction} 
\alias{pr.conditional} 
\alias{pr.marginal} 
\alias{print.cin} 
\alias{slashTerms}


\title{ Undocumented and Internally Used Functions}
\description{
  Those currently undocumented and internally used functions are aliased
  to this help file.
}


\details{
  In the \pkg{prLogistic} package there are currently many
  objects which are currently internal and/or
  undocumented. The help file suppresses the warnings when the package is
  'CHECK'ed.

}
\value{
  Each objects/methodsmay or may not have its own individual value.
  These will be documented over time.

}


\author{Raydonal Ospina, Department of Statistics, Federal University of Pernambuco, Brazil \cr (raydonal@de.ufpe.br) \cr 
Leila D. Amorim, Department of Statistics, Federal University of Bahia, Brazil \cr (leiladen@ufba.br).
}



\keyword{regression}
\keyword{internal}







