% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-prLogistic.R
\name{summary.prLogistic}
\alias{summary.prLogistic}
\title{Summarise a prLogistic object}
\usage{
\method{summary}{prLogistic}(object, ...)
}
\arguments{
\item{object}{A \code{prLogistic} object.}

\item{...}{Currently ignored.}
}
\description{
Summarise a prLogistic object
}
