% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-prLogistic.R
\name{print.prLogistic}
\alias{print.prLogistic}
\title{Print a prLogistic object}
\usage{
\method{print}{prLogistic}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{prLogistic} object.}

\item{digits}{Number of significant digits (default 4).}

\item{...}{Currently ignored.}
}
\description{
Print a prLogistic object
}
